Clean samples and add method docs
