+++ src/Ether.Network/Exceptions/EtherException.cs
@@ -2,13 +2,25 @@
 
 namespace Ether.Network.Exceptions
 {
+    /// <summary>
+    /// Represents a generic Ether.Network exception.
+    /// </summary>
     public class EtherException : Exception
     {
+        /// <summary>
+        /// Creates a new <see cref="EtherException"/>.
+        /// </summary>
+        /// <param name="message"></param>
         public EtherException(string message)
             : this(message, null)
         {
         }
 
+        /// <summary>
+        /// Creates a new <see cref="EtherException"/> with an inner exception.
+        /// </summary>
+        /// <param name="message"></param>
+        /// <param name="innerException"></param>
         public EtherException(string message, Exception innerException)
             : base(message, innerException)
         {
